Changed the broadcasting approach
Now the microphone broadcast uses a more compatible way to run, without the need of the jnaudiostream library. The raw microphone data is send over to the server via socke.io, but is converted using ffmpeg and statically served to the listener. The listener plays it inside a HTML audio tag, without the need to use Media Source Extentions that aren't compatible to the Apple mobile ecosystem.